move tabular entities in one place
