+++ src/main/java/org/gridsuite/modification/server/entities/tabular/TabularBaseEntity.java
@@ -4,7 +4,7 @@
   License, v. 2.0. If a copy of the MPL was not distributed with this
   file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.gridsuite.modification.server.entities;
+package org.gridsuite.modification.server.entities.tabular;
 
 import jakarta.persistence.CascadeType;
 import jakarta.persistence.FetchType;
@@ -16,6 +16,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.gridsuite.modification.dto.ModificationInfos;
 import org.gridsuite.modification.dto.TabularBaseInfos;
+import org.gridsuite.modification.server.entities.ModificationEntity;
 
 import java.util.List;
