+++ src/main/java/org/gridsuite/modification/server/entities/tabular/TabularModificationEntity.java
@@ -4,7 +4,7 @@
   License, v. 2.0. If a copy of the MPL was not distributed with this
   file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
-package org.gridsuite.modification.server.entities;
+package org.gridsuite.modification.server.entities.tabular;
 
 import jakarta.persistence.*;
 import lombok.Getter;
@@ -17,6 +17,7 @@ import java.util.stream.Collectors;
 
 import org.gridsuite.modification.ModificationType;
 import org.gridsuite.modification.dto.*;
+import org.gridsuite.modification.server.entities.ModificationEntity;
 import org.springframework.util.CollectionUtils;
 
 /**
